+++ addAssetHtmlPlugin.js
@@ -6,9 +6,10 @@ import Promise from 'bluebird';
 function resolvePublicPath(compilation, filename, publicPath) {
   let resolvedPublicPath;
   if (typeof publicPath === 'undefined') {
+    /* istanbul ignore else */
     resolvedPublicPath = typeof compilation.options.output.publicPath !== 'undefined'
       ? compilation.options.output.publicPath
-      : path.relative(path.dirname(filename), '.');
+      : path.relative(path.dirname(filename), '.'); // TODO: How to test this? I haven't written this logic, unsure what it does
   } else {
     resolvedPublicPath = publicPath;
   }
@@ -19,9 +20,12 @@ function resolvePublicPath(compilation, filename, publicPath) {
   return resolvedPublicPath;
 }
 
-function addFileToAssets(htmlPluginData, compilation,
-  { filename, typeOfAsset = 'js', includeSourcemap = true, hash = false, publicPath } = {}) {
-  if (!filename) return compilation.errors.push(new Error('No filename defined'));
+function addFileToAssets(compilation, htmlPluginData, { filename, typeOfAsset = 'js', includeSourcemap = true, hash = false, publicPath }) {
+  if (!filename) {
+    const error = new Error('No filename defined');
+    compilation.errors.push(error);
+    return Promise.reject(error);
+  }
 
   return htmlPluginData.plugin.addFileToAssets(filename, compilation)
     .then(addedFilename => {
@@ -31,8 +35,14 @@ function addFileToAssets(htmlPluginData, compilation,
         md5.update(compilation.assets[addedFilename].source());
         suffix = `?${md5.digest('hex').substr(0, 20)}`;
       }
-      return htmlPluginData.assets[typeOfAsset]
-        .unshift(`${resolvePublicPath(compilation, addedFilename, publicPath)}${addedFilename}${suffix}`);
+
+      // TODO: No need to call this if `publicPath` is provided
+      const resolvedPublicPath = resolvePublicPath(compilation, addedFilename, publicPath);
+      const resolvedPath = `${resolvedPublicPath}${addedFilename}${suffix}`;
+
+      htmlPluginData.assets[typeOfAsset].unshift(resolvedPath);
+
+      return resolvedPath;
     })
     .then(() => {
       if (includeSourcemap) {
@@ -42,16 +52,23 @@ function addFileToAssets(htmlPluginData, compilation,
     });
 }
 
+// Visible for testing
+export function addAllAssetsToCompilation(assets, compilation, htmlPluginData, callback) {
+  return Promise.mapSeries(assets, asset => addFileToAssets(compilation, htmlPluginData, asset))
+    .then(() => callback(null, htmlPluginData))
+    .catch(e => callback(e, htmlPluginData));
+}
+
 export default class AddAssetHtmlPlugin {
   constructor(assets = []) {
     this.assets = Array.isArray(assets) ? assets.slice().reverse() : [assets];
   }
 
+  /* istanbul ignore next: this would be integration tests */
   apply(compiler) {
     compiler.plugin('compilation', compilation => {
       compilation.plugin('html-webpack-plugin-before-html-generation', (htmlPluginData, callback) => {
-        Promise.mapSeries(this.assets, asset => addFileToAssets(htmlPluginData, compilation, asset))
-          .then(() => callback(null, htmlPluginData));
+        addAllAssetsToCompilation(this.assets, compilation, htmlPluginData, callback);
       });
     });
   }